flatpak-spawn: Try to normalize exposed paths to be in ~/.var/app
The Flatpak portal will reject paths that do not resolve to the same
thing in the container and on the host system. For example,

    --sandbox-expose-path="$HOME/.config/my/file"

will not work, even if

    --sandbox-expose-path="$HOME/.var/app/com.example.App/config/my/file"

would.

Existing apps like GNOME Web (Epiphany) get away with this because the
files they are exposing are below $XDG_CONFIG_HOME, $XDG_DATA_HOME or
$XDG_CACHE_HOME, which take values below ~/.var/app, but the same is
not necessarily going to be true for apps like Steam that use --persist.

Signed-off-by: Simon McVittie <smcv@collabora.com>